# contract/openapi.yaml
openapi: 3.1.0
info:
  title: Kestrel AI backend API
  version: 1.0.0
  description: |
    Contract between the Tauri/React frontend and the FastAPI sidecar. This file is the
    source of truth (spec section 2). The backend validates responses against it in tests,
    the TypeScript client is generated from it, and Prism serves it as the mock server.

    Conventions:
    - Every path below carries the `/api/v1` prefix. JSON bodies, UTF-8.
    - Every request carries the per-launch token, either as `Authorization: Bearer <token>`
      or as a `token` query parameter (needed for `<img>` sources and websocket connects).
    - Ids are UUID strings. Timestamps are RFC 3339 strings in UTC.
    - Boxes are in image pixels: `x, y` top-left corner, `w, h` size.
    - Every list endpoint paginates with `limit` and `cursor` and returns `next_cursor`
      (`null` when there is no further page).
    - Every error uses the `Error` envelope `{error: {code, message, details}}`.
    - Long work returns a `Job`; progress arrives over the websocket described in
      `x-websocket` and can be polled through the `jobs` resource.
servers:
  - url: http://127.0.0.1:8765
    description: development backend origin (the packaged app uses a random free port)
  - url: http://127.0.0.1:4010
    description: Prism mock server
security:
  - bearerAuth: []
  - tokenQuery: []
...
paths:
...
  /api/v1/projects/{projectId}:
    parameters:
      - $ref: "#/components/parameters/projectId"
    get:
      tags: [projects]
      operationId: getProject
      responses:
        "200":
          description: the project
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
        default: { $ref: "#/components/responses/Error" }
    patch:
      tags: [projects]
      operationId: updateProject
      summary: Update name, pre-annotation model or import defaults (project settings screen).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProjectUpdate" }
      responses:
        "200":
          description: updated project
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
        default: { $ref: "#/components/responses/Error" }
    delete:
      tags: [projects]
      operationId: forgetProject
      summary: Remove the project from the recent list. The project folder and everything in it stay on disk; "Open folder" brings it back.
      responses:
        "204":
          description: forgotten
        default: { $ref: "#/components/responses/Error" }
...
  /api/v1/projects/{projectId}/train:
    parameters:
      - $ref: "#/components/parameters/projectId"
    post:
      tags: [library]
      operationId: trainModel
      summary: |
        Start a training job in a training project. `base_model_id` is a library model id. When
        the job succeeds the trained weights are registered in the library (`origin: trained`)
        and `job.result.model_id` is the new library model id.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TrainRequest" }
      responses:
        "202":
          description: training job queued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobRef" }
        "404":
          description: the project, the dataset or the base library model does not exist (`code` is `not_found`)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: "the project is not a training project (`code` is `wrong_project_kind`, details `{kind, allowed}`), or the base model's weights file is missing (`code` is `model_unavailable`)"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422":
          description: the request body is invalid (`code` is `validation_error`)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503": { $ref: "#/components/responses/LibraryUnavailable" }
        default: { $ref: "#/components/responses/Error" }

  # ------------------------------------------------------------ model adoption
...
  /api/v1/library/status:
    get:
      tags: [library]
      operationId: getLibraryStatus
      summary: Whether the app-wide model library opened at startup, and where it lives. Never answers 503.
      responses:
        "200":
          description: library status
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LibraryStatus" }
        default: { $ref: "#/components/responses/Error" }
  /api/v1/library/models:
    get:
      tags: [library]
      operationId: listLibraryModels
      summary: Every model in the library, newest first.
      parameters:
        - name: task
          in: query
          schema: { type: string, enum: [detect, obb] }
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/cursor"
      responses:
        "200":
          description: library models
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LibraryModelPage" }
        "503": { $ref: "#/components/responses/LibraryUnavailable" }
        default: { $ref: "#/components/responses/Error" }
  /api/v1/library/models/import:
    post:
      tags: [library]
      operationId: importLibraryModel
      summary: |
        Import a `.pt` weights file into the library through a `library_import` job: hash, copy,
        load-check, read the task and class names, register. A file already in the library (same
        sha256) fails the job with a message naming the existing model. The source file is only read.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LibraryModelImport" }
      responses:
        "202":
          description: "import job queued; when it succeeds its `result` is `{model_id}`"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobRef" }
        "404":
          description: "`weights_path` is not absolute, not a .pt file, or does not exist (`code` is `not_found`)"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503": { $ref: "#/components/responses/LibraryUnavailable" }
        default: { $ref: "#/components/responses/Error" }
  /api/v1/library/models/{modelId}:
    parameters:
      - $ref: "#/components/parameters/modelId"
    get:
      tags: [library]
      operationId: getLibraryModel
      responses:
        "200":
          description: the library model
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LibraryModel" }
        "404":
          description: no such model (`code` is `not_found`)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503": { $ref: "#/components/responses/LibraryUnavailable" }
        default: { $ref: "#/components/responses/Error" }
    patch:
      tags: [library]
      operationId: updateLibraryModel
      summary: Rename the model or change its notes, supplier, class aliases or training scale.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LibraryModelPatch" }
      responses:
        "200":
          description: the updated library model
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LibraryModel" }
        "404":
          description: no such model (`code` is `not_found`)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503": { $ref: "#/components/responses/LibraryUnavailable" }
        default: { $ref: "#/components/responses/Error" }
    delete:
      tags: [library]
      operationId: deleteLibraryModel
      summary: |
        Remove the model and its folder from the library. Past runs keep their own copy of the
        model's name and classes and stay readable; check `/usage` first to warn the operator.
      responses:
        "204":
          description: deleted
        "404":
          description: no such model (`code` is `not_found`)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503": { $ref: "#/components/responses/LibraryUnavailable" }
        default: { $ref: "#/components/responses/Error" }
  /api/v1/library/models/{modelId}/gsd-estimate:
    parameters:
      - $ref: "#/components/parameters/modelId"
    get:
      tags: [library]
      operationId: getLibraryModelGsdEstimate
      summary: |
        The scale the model was trained at, derived from the dataset it was trained on, with the
        evidence for it. A bounded read: the median of an altitude column already in the project
        database plus at most eight EXIF headers, so it is a plain request and not a job.
      responses:
        "200":
          description: the derived training scale and its evidence
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ModelGsdEstimate" }
        "404":
          description: |
            no such model, or nothing can be measured - the model names no dataset, its originating
            project folder no longer opens, or its imagery carries no usable camera EXIF
            (`code` is `not_found`)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503": { $ref: "#/components/responses/LibraryUnavailable" }
        default: { $ref: "#/components/responses/Error" }
  /api/v1/library/models/{modelId}/usage:
    parameters:
      - $ref: "#/components/parameters/modelId"
    get:
      tags: [library]
      operationId: getLibraryModelUsage
      summary: Recently opened projects that use the model (pre-annotation, detection runs, map runs). Projects not on the recent list are not scanned.
      responses:
        "200":
          description: the projects that use the model
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ModelUsage" }
        "404":
          description: no such model (`code` is `not_found`)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503": { $ref: "#/components/responses/LibraryUnavailable" }
        default: { $ref: "#/components/responses/Error" }
  /api/v1/library/models/{modelId}/artifacts/{artifact}:
    parameters:
      - $ref: "#/components/parameters/modelId"
      - name: artifact
        in: path
        required: true
        schema:
          type: string
          enum: [results_csv, confusion_matrix, pr_curve]
    get:
      tags: [library]
      operationId: getLibraryModelArtifact
      summary: A training artifact file (results.csv, confusion matrix PNG, PR curve PNG) kept with the library model.
      responses:
        "200":
          description: the file
          content:
            image/png:
              schema: { type: string, format: binary }
            text/csv:
              schema: { type: string }
        "404":
          description: no such model, or the model has no such artifact (`code` is `not_found`)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503": { $ref: "#/components/responses/LibraryUnavailable" }
        default: { $ref: "#/components/responses/Error" }
  /api/v1/library/models/{modelId}/export:
    parameters:
      - $ref: "#/components/parameters/modelId"
    post:
      tags: [library]
      operationId: exportLibraryModel
      summary: Export to ONNX or TensorRT through a `library_export` job; the path lands in `exports[format]` of the library model.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ExportRequest" }
      responses:
        "202":
          description: "export job queued; when it succeeds its `result` is `{format, path}`"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobRef" }
        "404":
          description: no such model (`code` is `not_found`)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503": { $ref: "#/components/responses/LibraryUnavailable" }
        default: { $ref: "#/components/responses/Error" }
  /api/v1/library/starters/{key}/acquire:
    parameters:
      - name: key
        in: path
        required: true
        schema: { $ref: "#/components/schemas/StarterModelKey" }
    post:
      tags: [library]
      operationId: acquireStarterModel
      summary: Add one supported YOLO starter to the library through a `library_starter` job.
      description: Reuses bundled or cached weights; otherwise downloads the chosen asset. Progress and cancellation use the library jobs resource. The successful `job.result.model_id` identifies the library model.
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StarterAcquire" }
      responses:
        "202":
          description: starter acquisition queued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobRef" }
        "404":
          description: no such starter model (`code` is `not_found`)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503": { $ref: "#/components/responses/LibraryUnavailable" }
        default: { $ref: "#/components/responses/Error" }
  /api/v1/library/jobs:
    get:
      tags: [library]
      operationId: listLibraryJobs
      summary: Library jobs (import, export, starter), newest first. Their `project_id` is `library`.
      parameters:
        - name: state
          in: query
          schema: { $ref: "#/components/schemas/JobState" }
        - name: type
          in: query
          schema: { $ref: "#/components/schemas/JobType" }
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/cursor"
      responses:
        "200":
          description: library jobs, newest first
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobPage" }
        "503": { $ref: "#/components/responses/LibraryUnavailable" }
        default: { $ref: "#/components/responses/Error" }
  /api/v1/library/jobs/{jobId}:
    parameters:
      - $ref: "#/components/parameters/jobId"
    get:
      tags: [library]
      operationId: getLibraryJob
      responses:
        "200":
          description: the job
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Job" }
        "404":
          description: no such job (`code` is `not_found`)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503": { $ref: "#/components/responses/LibraryUnavailable" }
        default: { $ref: "#/components/responses/Error" }
  /api/v1/library/jobs/{jobId}/log:
    parameters:
      - $ref: "#/components/parameters/jobId"
    get:
      tags: [library]
      operationId: getLibraryJobLog
      parameters:
        - name: tail
          in: query
          schema: { type: integer, minimum: 1, maximum: 10000, default: 200 }
      responses:
        "200":
          description: last lines of the job log
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobLog" }
        "404":
          description: no such job (`code` is `not_found`)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503": { $ref: "#/components/responses/LibraryUnavailable" }
        default: { $ref: "#/components/responses/Error" }
  /api/v1/library/jobs/{jobId}/cancel:
    parameters:
      - $ref: "#/components/parameters/jobId"
    post:
      tags: [library]
      operationId: cancelLibraryJob
      summary: Request cancellation. The state becomes `cancelled` once the worker notices; finished jobs are returned unchanged.
      responses:
        "200":
          description: the job after the request
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Job" }
        "404":
          description: no such job (`code` is `not_found`)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503": { $ref: "#/components/responses/LibraryUnavailable" }
        default: { $ref: "#/components/responses/Error" }
...
  /api/v1/projects/{projectId}/jobs:
    parameters:
      - $ref: "#/components/parameters/projectId"
    get:
      tags: [jobs]
      operationId: listJobs
      parameters:
        - name: state
          in: query
          schema: { $ref: "#/components/schemas/JobState" }
        - name: type
          in: query
          schema: { $ref: "#/components/schemas/JobType" }
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/cursor"
      responses:
        "200":
          description: jobs, newest first
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobPage" }
        default: { $ref: "#/components/responses/Error" }
  /api/v1/projects/{projectId}/jobs/{jobId}:
    parameters:
      - $ref: "#/components/parameters/projectId"
      - $ref: "#/components/parameters/jobId"
    get:
      tags: [jobs]
      operationId: getJob
      responses:
        "200":
          description: the job
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Job" }
        default: { $ref: "#/components/responses/Error" }
  /api/v1/projects/{projectId}/jobs/{jobId}/cancel:
    parameters:
      - $ref: "#/components/parameters/projectId"
      - $ref: "#/components/parameters/jobId"
    post:
      tags: [jobs]
      operationId: cancelJob
      summary: Request cancellation. The state becomes `cancelled` once the worker notices; finished jobs are returned unchanged.
      responses:
        "200":
          description: the job after the request
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Job" }
        default: { $ref: "#/components/responses/Error" }
  /api/v1/projects/{projectId}/jobs/{jobId}/log:
    parameters:
      - $ref: "#/components/parameters/projectId"
      - $ref: "#/components/parameters/jobId"
    get:
      tags: [jobs]
      operationId: getJobLog
      parameters:
        - name: tail
          in: query
          schema: { type: integer, minimum: 1, maximum: 10000, default: 200 }
      responses:
        "200":
          description: last lines of the job log
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobLog" }
        default: { $ref: "#/components/responses/Error" }

  # ----------------------------------------------------------------- project agent
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: the per-launch token passed to the sidecar in `APP_TOKEN`
    tokenQuery:
      type: apiKey
      in: query
      name: token
      description: same token as a query parameter, for image URLs and the websocket
  parameters:
    projectId:
      name: projectId
      in: path
      required: true
      schema: { type: string }
...
  responses:
    Error:
      description: error envelope
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    LibraryUnavailable:
      description: "the model library could not be opened at startup (`code` is `library_unavailable`); `GET /library/status` has the reason"
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error:
              code: library_unavailable
              message: "The model library could not be opened."
              details: {}
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, details]
          properties:
            code:
              type: string
              description: |
                machine-readable: unauthorized, not_found, validation_error, already_exists,
                class_in_use, conflict, not_implemented, provider_error, internal_error,
                wrong_project_kind (409: the operation does not belong to this kind of project;
                details `{kind, allowed}`), library_unavailable (503: the model library could not
                be opened), model_unavailable (409: the library model's weights file is missing),
                unmapped_classes (422: the run's model has classes with no project class and no
                remembered mapping; details `{model_id, unmapped}`), unsupported_point_cloud,
                insufficient_memory and insufficient_disk (422: a point cloud cannot be read or
                admitted), link_needs_coordinates, no_overlap and crs_already_set (422: a point
                cloud patch), invalid_epsg (422: an EPSG code pyproj does not know),
                wrong_point_count, needs_projected_crs, vertical_span_too_small and
                measurement_limit (422: a point-cloud measurement), octree_missing (404: a ready
                cloud's display-copy file is missing on disk), grid_too_large (422: a surface
                grid over the cell ceiling), unsupported_crs, source_missing and
                invalid_build_request (422: a surface build), no_coordinates (422: an ortho
                underlay without a CRS), invalid_geometry and invalid_base (422: a volume
                measurement's polygon or base), job_running (409: the resource's job is queued
                or running), not_ready (409: the resource has not finished importing),
                range_not_satisfiable (416)
            message: { type: string }
            details:
              type: object
              additionalProperties: true
      example:
        error: { code: not_found, message: "project 7f1c… not found", details: {} }
...
    Project:
      type: object
      required: [id, name, folder, kind, classes, preannotation_model_id, import_defaults, schema_version, created_at]
      properties:
        id: { type: string }
        name: { type: string }
        folder: { type: string, description: absolute path of the project folder }
        kind: { $ref: "#/components/schemas/ProjectKind" }
        classes:
          type: array
          items: { $ref: "#/components/schemas/ClassDef" }
        preannotation_model_id: { type: [string, "null"], description: a library model id }
        import_defaults: { $ref: "#/components/schemas/ImportSettings" }
        schema_version: { type: integer }
        created_at: { type: string, format: date-time }
      example:
        id: "7f1c2e3a-1111-4000-8000-000000000001"
        name: Ahmadia
        folder: "E:\\Projects\\Ahmadia"
        kind: train
        classes:
          - { id: "c1a2b3c4-0000-4000-8000-000000000001", name: excavator, colour: "#f97316", hotkey: "1", order: 0 }
          - { id: "c1a2b3c4-0000-4000-8000-000000000002", name: wheel_loader, colour: "#eab308", hotkey: "2", order: 1 }
          - { id: "c1a2b3c4-0000-4000-8000-000000000003", name: bulldozer, colour: "#22c55e", hotkey: "3", order: 2 }
          - { id: "c1a2b3c4-0000-4000-8000-000000000004", name: dump_truck, colour: "#06b6d4", hotkey: "4", order: 3 }
          - { id: "c1a2b3c4-0000-4000-8000-000000000005", name: crane, colour: "#3b82f6", hotkey: "5", order: 4 }
          - { id: "c1a2b3c4-0000-4000-8000-000000000006", name: concrete_mixer, colour: "#a855f7", hotkey: "6", order: 5 }
          - { id: "c1a2b3c4-0000-4000-8000-000000000007", name: roller, colour: "#ec4899", hotkey: "7", order: 6 }
          - { id: "c1a2b3c4-0000-4000-8000-000000000008", name: backhoe, colour: "#ef4444", hotkey: "8", order: 7 }
        preannotation_model_id: "m0000000-2222-4000-8000-000000000001"
        import_defaults: { max_side: 4000, quality: 95, dedupe_threshold: 4, group_regex: "^(?P<camera>[A-Za-z0-9-]+)_(?P<flight>\\d+)_(?P<frame>\\d+)" }
        schema_version: 1
        created_at: "2026-09-17T10:00:00Z"
...
    ProjectUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        preannotation_model_id: { type: [string, "null"], description: a library model id }
        import_defaults: { $ref: "#/components/schemas/ImportSettings" }
      example: { name: Ahmadia, preannotation_model_id: "m0000000-2222-4000-8000-000000000001" }
...
    TrainRequest:
      type: object
      required: [name, dataset_id, base_model_id]
      properties:
        name: { type: string, minLength: 1 }
        dataset_id: { type: string }
        base_model_id: { type: string, description: "a library model id (any library model, including starter COCO weights)" }
        epochs: { type: integer, minimum: 1, maximum: 1000, default: 50 }
        imgsz: { type: integer, minimum: 320, maximum: 4096, default: 1280 }
        batch: { type: [integer, "null"], description: null = auto }
        patience: { type: integer, minimum: 0, default: 50 }
        augmentation:
          type: string
          enum: [default, aerial]
          default: default
        device: { type: string, default: "0", description: "\"0\" for the first GPU, \"cpu\" otherwise" }
      example: { name: ahmadia-v1-n, dataset_id: "d0000000-7777-4000-8000-000000000001", base_model_id: "m0000000-2222-4000-8000-000000000002", epochs: 3, imgsz: 1280, batch: null, patience: 50, augmentation: aerial, device: "0" }
...
    JobRef:
      type: object
      required: [job]
      properties:
        job: { $ref: "#/components/schemas/Job" }
...